// src/app/models/producto.ts
export class producto {
    idprodinv: number;
    categoria: string;
    dateadd: string;
    datemod: string;
    estado: string;
    existencia: number;
    idsucursal: number;
    iduseradd: number;
    idusermod: number;
    nombre: string;
    preciocosto: number;
    preciooferta: number;
    precioregular: number;
    serie: string;
    cantidad: number;
    precioTotal: number;
}